+++ bootstrap/providers.php
@@ -4,4 +4,5 @@ use App\Providers\AppServiceProvider;
 
 return [
     AppServiceProvider::class,
+    Spatie\Backup\BackupServiceProvider::class,
 ];